+++ lib/recording_studio_ai/models.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require "recording_studio_ai/models/definition"
+require "recording_studio_ai/models/registry"